back-server/requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
cryptography==40.0.2
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==4.0.0
Django==3.2.18
django-allauth==0.52.0
django-cors-headers==3.14.0
django-extensions==3.2.1
djangorestframework==3.14.0
executing==1.2.0
idna==3.4
ipython==8.12.0
jedi==0.18.2
matplotlib-inline==0.1.6
oauthlib==3.2.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.38
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==2.0.2
wcwidth==0.2.6
